requirements_ml.txt
torch>=2.0
torchvision>=0.15
facenet-pytorch>=2.5
scikit-learn>=1.3
numpy>=1.24
matplotlib>=3.7
Pillow>=9.0
